Id mudado de long para int Adicionado tela provisória de login, funcionando

+++ src/java/com/ServeletCliente/AlterarCliente.java
@@ -23,7 +23,7 @@ import javax.servlet.http.HttpServletResponse;
 public class AlterarCliente extends HttpServlet {
     @Override
     protected void service (HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException{ 
-        Long id = Long.parseLong(req.getParameter("id"));
+        int id = Integer.parseInt(req.getParameter("id"));
            
         ClienteModel dao = new ClienteModel();
         ClienteBean cli = dao.buscarId(id);
